+++ R/get_data.R
@@ -833,6 +833,38 @@ get_data.stanmvreg <- function(x, ...) {
 
 
 #' @export
+get_data.averaging <- function(x, ...) {
+  ml <- attributes(x)$modelList
+  if (is.null(ml)) {
+    warning("Can't retrieve data. Please use 'fit = TRUE' in 'model.avg()'.", call. = FALSE)
+    return(NULL)
+  }
+
+  mf <- tryCatch(
+    {
+      Reduce(function(x, y) merge(x, y, all = TRUE, sort = FALSE), lapply(ml, stats::model.frame))
+    },
+    error = function(x) {
+      NULL
+    }
+  )
+
+  if (is.null(mf)) {
+    mf <- tryCatch(
+      {
+        .get_data_from_env(x)[, find_variables(x, flatten = TRUE), drop = FALSE]
+      },
+      error = function(x) {
+        NULL
+      }
+    )
+  }
+
+  .prepare_get_data(x, mf)
+}
+
+
+#' @export
 get_data.Arima <- function(x, ...) {
   # first try, parent frame
   dat <- tryCatch(